+++ src/HDFJavaUtils/ObjectOutputStream.java
@@ -778,7 +778,7 @@ public class ObjectOutputStream {
 	private static long[] intToLongArr(int[] intArr) {
 		long[] longArr = new long[intArr.length];
 		for (int i = 0; i < intArr.length; i++) {
-			longArr[i] = (long) intArr[i];
+			longArr[i] = Long.valueOf(intArr[i]).intValue();
 		}
 		return longArr;
 	}
